requirements.txt
awsglue==4.0
spark==3.3.0
python==3.10
